# sql/catalog: restore fast-path in FullIndexColumnIDs