@postamar 2022-01-10
> Added a comment for now, but precomputing these slices sounds like a great idea.

I went ahead and created https://github.com/cockroachdb/cockroach/pull/74638